feat: implement todo01 — async database writes
Move telemetry DB writes off the per-client recv thread onto a dedicated
worker via a bounded write queue with drop-oldest backpressure. The recv
thread now returns immediately from enqueue(), so a slow or stalled DB
cannot delay safety-critical command delivery.

+++ src/db.cpp
@@ -48,7 +48,7 @@ flight_safety_system::server::db_connection::recordSearchStatus(uint64_t asset_i
 }
 
 void
-flight_safety_system::server::db_connection::recordPosition(uint64_t asset_id, double latitude, double longitude, uint16_t altitude)
+flight_safety_system::server::db_connection::recordPosition(uint64_t asset_id, double latitude, double longitude, uint32_t altitude)
 {
     std::lock_guard<std::mutex> guard(this->db_lock);
     db_position_create_entry(asset_id, latitude, longitude, altitude);
